+++ beams/tree_config/__init__.py
@@ -35,13 +35,16 @@ __all__ = [
     "SetPVActionItem",
     "IncPVActionItem",
     "CheckAndDoItem",
-    "UseCheckConditionItem"
+    "UseCheckConditionItem",
+    "IncPVActionItem",
+    "SetPVActionItem"
 ]
 from .base import (BaseItem, BehaviorTreeItem, ExternalItem, PVTarget, Target,
                    ValueTarget)
 from .condition import (BaseConditionItem, BinaryConditionItem,
                         ConditionOperator, DummyConditionItem,
                         ThresholdConditionItem)
+from .action import (IncPVActionItem, SetPVActionItem)
 from .py_trees import (BlackboardToStatusItem,
                        CheckBlackboardVariableExistsItem,
                        CheckBlackboardVariableValueItem, DummyItem,
@@ -51,8 +54,8 @@ from .py_trees import (BlackboardToStatusItem,
                        UnsetBlackboardVariableItem,
                        WaitForBlackboardVariableItem,
                        WaitForBlackboardVariableValueItem)
-from .tree_config import (BaseSequenceItem, CheckAndDoItem, IncPVActionItem,
+from .tree_config import (BaseSequenceItem, CheckAndDoItem,
                           ParallelItem, ParallelMode, SelectorItem,
-                          SequenceConditionItem, SequenceItem, SetPVActionItem,
+                          SequenceConditionItem, SequenceItem,
                           UseCheckConditionItem, get_tree_from_path,
                           save_tree_item_to_path)
