+++ beams/tree_config/tree_config.py
@@ -6,19 +6,19 @@ from copy import copy
 from dataclasses import dataclass, field
 from enum import Enum
 from pathlib import Path
-from typing import Any, List, Union
+from typing import List, Union
 
 import py_trees
 from apischema import deserialize, serialize
-from epics import caget, caput
+
 from py_trees.common import ParallelPolicy
 from py_trees.composites import Parallel, Selector, Sequence
 
-from beams.behavior_tree.ActionNode import ActionNode, wrapped_action_work
 from beams.behavior_tree.CheckAndDo import CheckAndDo
 from beams.serialization import as_tagged_union
 from beams.tree_config.base import BaseItem, BehaviorTreeItem
 from beams.tree_config.condition import BaseConditionItem, DummyConditionItem
+from beams.tree_config.action import SetPVActionItem, IncPVActionItem
 from beams.typing_helper import Evaluatable
 
 logger = logging.getLogger(__name__)
@@ -150,85 +150,6 @@ class SequenceConditionItem(BaseSequenceItem, BaseConditionItem):
             return ok
 
         return cond_func
-
-
-@dataclass
-class SetPVActionItem(BaseItem):
-    pv: str = ""
-    value: Any = 1
-    loop_period_sec: float = 1.0
-
-    termination_check: BaseConditionItem = field(default_factory=DummyConditionItem)
-
-    def get_tree(self) -> ActionNode:
-
-        @wrapped_action_work(self.loop_period_sec)
-        def work_func(comp_condition: Evaluatable) -> py_trees.common.Status:
-            try:
-                # Set to running
-                value = caget(self.pv)  # double caget, this is uneeded as currently the comp_condition has caget baked in
-
-                if comp_condition():
-                    return py_trees.common.Status.SUCCESS
-                logger.debug(f"{self.name}: Value is {value}")
-
-                # specific caput logic to SetPVActionItem
-                caput(self.pv, self.value)
-                return py_trees.common.Status.RUNNING
-            except Exception as ex:
-                logger.warning(f"{self.name}: work failed: {ex}")
-                return py_trees.common.Status.FAILURE
-
-        comp_cond = self.termination_check.get_condition_function()
-
-        node = ActionNode(
-            name=self.name,
-            work_func=work_func,
-            completion_condition=comp_cond,
-        )
-
-        return node
-
-
-@dataclass
-class IncPVActionItem(BaseItem):
-    pv: str = ""
-    increment: float = 1
-    loop_period_sec: float = 1.0
-
-    termination_check: BaseConditionItem = field(default_factory=DummyConditionItem)
-
-    def get_tree(self) -> ActionNode:
-
-        @wrapped_action_work(self.loop_period_sec)
-        def work_func(comp_condition: Evaluatable) -> py_trees.common.Status:
-            """
-            To be run inside of a while loop
-            Action node should take care of logging, reporting status
-            """
-            try:
-                value = caget(self.pv)
-
-                logging.debug(f"(wf) {self.name}: Value is {value}")
-                if comp_condition():
-                    return py_trees.common.Status.SUCCESS
-
-                # specific caput logic to IncPVActionItem
-                caput(self.pv, value + self.increment)
-                return py_trees.common.Status.RUNNING
-            except Exception as ex:
-                logger.warning(f"{self.name}: work failed: {ex}")
-                return py_trees.common.Status.FAILURE
-
-        comp_cond = self.termination_check.get_condition_function()
-
-        node = ActionNode(
-            name=self.name,
-            work_func=work_func,
-            completion_condition=comp_cond,
-        )
-
-        return node
 
 
 @dataclass
